+++ src/lib/core/CHIPTLVReader.cpp
@@ -266,7 +266,7 @@ CHIP_ERROR TLVReader::Get(ByteSpan & v)
     return CHIP_NO_ERROR;
 }
 
-CHIP_ERROR TLVReader::GetBytes(uint8_t * buf, uint32_t bufSize)
+CHIP_ERROR TLVReader::GetBytes(uint8_t * buf, size_t bufSize)
 {
     if (!TLVTypeIsString(ElementType()))
         return CHIP_ERROR_WRONG_TLV_TYPE;
@@ -283,7 +283,7 @@ CHIP_ERROR TLVReader::GetBytes(uint8_t * buf, uint32_t bufSize)
     return CHIP_NO_ERROR;
 }
 
-CHIP_ERROR TLVReader::GetString(char * buf, uint32_t bufSize)
+CHIP_ERROR TLVReader::GetString(char * buf, size_t bufSize)
 {
     if (!TLVTypeIsString(ElementType()))
         return CHIP_ERROR_WRONG_TLV_TYPE;
